WIP : still shokhm AI

+++ parsian_ai/include/parsian_ai/roles/dynamicrole.h
@@ -30,7 +30,7 @@ private:
     SkillProperty(CRoleDynamic, double, Tolerance, tolerance);
     SkillProperty(CRoleDynamic, bool, Chip, chip);
     SkillProperty(CRoleDynamic, int, KickSpeed, kickSpeed);
-    SkillProperty(CRoleDynamic, float, ReceiveRadius, receiveRadius);
+    SkillProperty(CRoleDynamic, double, ReceiveRadius, receiveRadius);
     SkillProperty(CRoleDynamic, Vector2D, WaitPos, waitPos);
     SkillProperty(CRoleDynamic, bool, VeryFine, veryFine);
     SkillProperty(CRoleDynamic, bool, EmptySpot, emptySpot);
